SQL: use BIGINT insted of INT in osm related tables

+++ sql_scripts/create_street_references_table.sql
@@ -1,6 +1,6 @@
 CREATE table street_references (
-    ID INTEGER NOT NULL AUTO_INCREMENT,
-    street_id INTEGER,
-    section_id DOUBLE,
+    ID BIGINT(20) NOT NULL AUTO_INCREMENT,
+    street_id BIGINT(20),
+    section_id BIGINT(20),
     PRIMARY KEY(ID)
 );
