+++ sql_scripts/create_streets_table.sql
@@ -1,5 +1,5 @@
 CREATE table streets (
-    street_id INTEGER NOT NULL,
+    street_id BIGINT(20) NOT NULL,
     name_of_street VARCHAR(200),
     street_of_budapest BOOLEAN,
     PRIMARY KEY(street_id)
